// visualization/src/generated_types/MirNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MirStmt } from "./MirStmt";

export type MirNode = { id: string, block: number, stmts: Array<MirStmt>, terminator: MirStmt, 
/**
 * DOT graph of the signature shape (from fn def signature), if this
 * block's terminator is a call.
 */
sig_shape_dot: string | null, 
/**
 * DOT graph of the call shape derived solely from call-site operand types.
 */
call_shape_dot: string | null, };