// visualization/src/generated_types/MirStmt.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MirStmtSpan } from "./MirStmtSpan";

export type MirStmt = { stmt: string, debug_stmt: string, span: MirStmtSpan, loans_invalidated_start: Array<string>, loans_invalidated_mid: Array<string>, borrows_in_scope_start: Array<string>, borrows_in_scope_mid: Array<string>, };